Import SVGs used in Layout page
